Integrate the Screen-Space Effects and NPE Sprite2D demos
Two upstream lab demos become registered application gates, measured on
SDL_GPU and Dawn.

Screen-Space Effects: the pinned screen-space task factories are recorded
at generation and lowered into both frame graphs, and the host UI companion
is loaded through one validator shared by generation and both capture
harnesses. The retained buttons take Chromium's button default font, which
also lowers the residual of every scene with toggle buttons (scene 4,
scene 22, Racer, Sandblox).

NPE Sprite2D: a node-particle set a pure-2D binding takes live is lowered
from its graph. The pinned block evaluators' build bodies are partially
evaluated at generation over the parsed graph, the closures they install
are translated through the pinned numeric translator, and the executed
pin's own build facts (installed slots, step count, settings, traversal
order) are checked against that evaluation. Both sprite passes fill a
loaded atlas's mip chain, the frame delta is carried as a double through
the frame clock, and the executed-URL and pending-let compiler contracts
cover the demo body.

Both demos meet their registered gates on both backends; NPE Sprite2D is
byte-exact. The simplify record covers both halves, and the full sweep
(258 scenes built, 257 measured within their gates) is green.

+++ src/lowering/physics-lowerer.ts
@@ -2678,9 +2678,8 @@ void fo_step_world(PhysicsWorld& world, double dt) {
     const double step_ms =
         world.fixed_delta_ms > 0.0
             ? world.fixed_delta_ms
-            : (world.scene != nullptr &&
-               static_cast<double>(world.scene->fixed_delta_ms) > 0.0
-                   ? static_cast<double>(world.scene->fixed_delta_ms)
+            : (world.scene != nullptr && world.scene->fixed_delta_ms > 0.0
+                   ? world.scene->fixed_delta_ms
                    : world.engine_delta_ms);
     if (!std::isfinite(step_ms) || step_ms <= 0.0) {
         return 0.0;
